feat: separate app files, update configs
