+++ .vscode/launch.json
@@ -9,7 +9,7 @@
       "type": "go",
       "request": "launch",
       "mode": "auto",
-      "program": "${workspaceFolder}/cmd/main.go",
+      "program": "${workspaceFolder}/cmd/grpcserver/main.go",
       "envFile": "${workspaceFolder}/dev.env"
     }
   ]
